requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
websocket-client==1.7.0
requests==2.31.0
windows-curses; sys_platform == "win32"
torch==2.3.1
torchvision==0.13.1
